requirements.txt
discord.py
loguru
python-dotenv
requests
rich
websocket-client[optional]